realtyna/requirements.txt
Django==4.1.1
psycopg2==2.9.3
python-decouple==3.6
djangorestframework==3.14.0
django-filter==22.1